% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_ad.tsvets.spec}
\alias{estimate_ad.tsvets.spec}
\alias{estimate_ad}
\title{Estimates a VETS model given a specification object using maximum likelihood and autodiff}
\usage{
\method{estimate_ad}{tsvets.spec}(
  object,
  solver = "nlminb",
  control = list(trace = 0, eval.max = 300, iter.max = 500),
  use_hessian = FALSE,
  power_iterations = 500,
  ...
)
}
\arguments{
\item{object}{An object of class tsvets.spec.}

\item{solver}{Only \dQuote{nlminb} currently supported.}

\item{control}{Solver control parameters.}

\item{use_hessian}{Whether to include the hessian in the calculation.}

\item{power_iterations}{number of iterations for the fast power iterations function
for finding the maximum eigenvalues during estimation.}

\item{...}{additional parameters passed to the estimation function}
}
\value{
An list of coefficients and other information.
}
\description{
Estimates a VETS model given a specification object using maximum likelihood and autodiff
}
\details{
This function is not expected to be used by itself but rather as a plugin
to be called from the estimate method of the tsvets package.
}
