% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvetsad-package.R
\docType{package}
\name{tsvetsad-package}
\alias{tsvetsad}
\alias{tsvetsad-package}
\title{tsvetsad: VETS models estimation using automatic differentiation}
\description{
Plugin Estimator for the tsvets package to use autodiff.
}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com}

}
\keyword{internal}
